+++ packages/terra-avatar/src/terra-dev-site/doc/example/avatar/AvatarImage.jsx
@@ -3,7 +3,7 @@ import Avatar from 'terra-avatar';
 import exampleAvatarImage from '../../../assets/150x150.jpg';
 
 const AvatarImage = () => (
-  <Avatar image={exampleAvatarImage} alt="Deep Space" />
+  <Avatar image={exampleAvatarImage} initials="JS" alt="Deep Space" />
 );
 
 export default AvatarImage;