+++ packages/terra-avatar/src/terra-dev-site/doc/example/avatar/AvatarUser.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Avatar from 'terra-avatar';
 
 const AvatarUser = () => (
-  <Avatar alt="The Last Airbender" />
+  <Avatar alt="The Last Airbender" initials="JS" />
 );
 
 export default AvatarUser;