+++ packages/terra-avatar/src/terra-dev-site/test/avatar/avatar/ImageAvatar.test.jsx
@@ -2,4 +2,4 @@ import React from 'react';
 import Avatar from '../../../../index';
 import exampleAvatarImage from '../../../assets/150x150.jpg';
 
-export default () => <Avatar image={exampleAvatarImage} alt="User" id="image-avatar" />;
+export default () => <Avatar image={exampleAvatarImage} alt="User" id="image-avatar" initials="JS" />;