+++ packages/terra-avatar/src/terra-dev-site/test/avatar/avatar/ImageGap.test.jsx
@@ -9,7 +9,7 @@ const cx = classNames.bind(styles);
 export default () => (
   <div>
     <div className={cx('avatar-wrapper')}>
-      <Avatar id="image-avatar" image={exampleAvatarImage} alt="placeholder" />
+      <Avatar id="image-avatar" image={exampleAvatarImage} alt="placeholder" initials="JS" />
     </div>
     <div className={cx('avatar-wrapper')}>
       <Facility id="facility-avatar" alt="facility" />