+++ packages/terra-avatar/src/terra-dev-site/test/avatar/avatar/IsDeceasedImageAvatar.test.jsx
@@ -2,4 +2,4 @@ import React from 'react';
 import Avatar from '../../../../index';
 import exampleAvatarImage from '../../../assets/150x150.jpg';
 
-export default () => <Avatar id="is-deceased-image-avatar" alt="user" image={exampleAvatarImage} isDeceased />;
+export default () => <Avatar id="is-deceased-image-avatar" alt="user" image={exampleAvatarImage} isDeceased initials="JS" />;