+++ packages/terra-avatar/src/terra-dev-site/test/avatar/avatar/UserAvatar.test.jsx
@@ -1,4 +1,4 @@
 import React from 'react';
 import Avatar from '../../../../index';
 
-export default () => <Avatar id="user-avatar" alt="user" color="one" />;
+export default () => <Avatar id="user-avatar" alt="user" color="one" initials="JS" />;