+++ packages/terra-avatar/tests/jest/Avatar.test.jsx
@@ -5,7 +5,7 @@ import exampleProfilePhoto from '../../src/terra-dev-site/assets/150x150.jpg';
 describe('Avatar', () => {
   // Snapshot Tests
   it('should render a default avatar', () => {
-    const avatar = <Avatar alt="user" />;
+    const avatar = <Avatar initials="JS" alt="user" />;
     const wrapper = render(avatar);
     expect(wrapper).toMatchSnapshot();
   });
@@ -22,14 +22,14 @@ describe('Avatar', () => {
     expect(wrapper).toMatchSnapshot();
   });
 
-  it('should render fallback user avatar when more than two initials are passed in', () => {
+  it('should render two initials avatar when more than two initials are passed in', () => {
     const avatar = <Avatar initials="JJJ" alt="user" />;
     const wrapper = render(avatar);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render image avatar when image is passed in', () => {
-    const avatar = <Avatar image={exampleProfilePhoto} alt="placeholder" />;
+    const avatar = <Avatar image={exampleProfilePhoto} alt="placeholder" initials="JS" />;
     const wrapper = render(avatar);
     expect(wrapper).toMatchSnapshot();
   });
@@ -41,13 +41,13 @@ describe('Avatar', () => {
   });
 
   it('should render fallback user avatar when invalid image is passed in without initials', () => {
-    const avatar = <Avatar image="https://path/to/invalid_image.jpg" alt="placeholder" />;
+    const avatar = <Avatar image="https://path/to/invalid_image.jpg" alt="placeholder" initials="JS" />;
     const wrapper = render(avatar);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an avatar with provided size', () => {
-    const avatar = <Avatar alt="user" size="5em" />;
+    const avatar = <Avatar alt="user" size="5em" initials="JS" />;
     const wrapper = render(avatar);
     expect(wrapper).toMatchSnapshot();
   });
@@ -59,31 +59,31 @@ describe('Avatar', () => {
   });
 
   it('should render an avatar with color variant one', () => {
-    const avatar = <Avatar alt="user" color="one" />;
+    const avatar = <Avatar alt="user" color="one" initials="JS" />;
     const wrapper = render(avatar);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an avatar with color variant netural', () => {
-    const avatar = <Avatar alt="user" color="neutral" />;
+    const avatar = <Avatar alt="user" color="neutral" initials="JS" />;
     const wrapper = render(avatar);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an avatar with an automated color variant, based on a hashValue', () => {
-    const avatar = <Avatar alt="user" hashValue="alternative hash" />;
+    const avatar = <Avatar alt="user" hashValue="alternative hash" initials="JS" />;
     const wrapper = render(avatar);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an isDeceased avatar', () => {
-    const avatar = <Avatar alt="user" isDeceased />;
+    const avatar = <Avatar alt="user" isDeceased initials="JS" />;
     const wrapper = render(avatar);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an avatar with custom props', () => {
-    const avatar = <Avatar alt="user" id="custom props avatar" />;
+    const avatar = <Avatar alt="user" id="custom props avatar" initials="JS" />;
     const wrapper = render(avatar);
     expect(wrapper).toMatchSnapshot();
   });