+++ packages/terra-avatar/tests/wdio/avatar-common-spec.js
@@ -31,12 +31,6 @@ Terra.describeViewports('Avatar', ['huge'], () => {
     Terra.it.validatesElement({ selector: '#invalid-image-avatar' });
   });
 
-  describe('Invalid Image Avatar Without Initials', () => {
-    before(() => browser.url('/#/raw/tests/terra-avatar/avatar/avatar/invalid-image-avatar-without-initials'));
-
-    Terra.it.validatesElement({ selector: '#invalid-image-avatar' });
-  });
-
   describe('Image Avatar Spacing', () => {
     before(() => browser.url('/#/raw/tests/terra-avatar/avatar/avatar/image-gap'));
     Terra.it.validatesElement();