+++ packages/terra-avatar/CHANGELOG.md
@@ -3,8 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
-### Changed 
-* Changed the fallback behavior of Avatar. Such that if the image fails to load when initials are provided, the avatar falls back to the initials display instead of default user icon.
+### Breaking Changes
+* Changed the fallback behavior of Avatar. Such that if the image fails to load, the avatar falls back to the initials display instead of default user icon.
 
 2.26.0 - (September 6, 2019)
 ------------------